+++ Integrations/JiraV2/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+API token parameter is now encrypted.
 
 ## [19.8.2] - 2019-08-22
 - Add support for remote application links.