+++ Integrations/JiraV2/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+
+
+## [19.10.2] - 2019-10-29
 API token parameter is now encrypted.
 
 ## [19.8.2] - 2019-08-22